refactor: 레포지토리 구조 수정
