+++ app/src/main/java/woowacourse/shopping/data/repository/cart/CartRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.repository
+package woowacourse.shopping.data.repository.cart
 
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.coroutines.flow.StateFlow
